+++ hpOneView/resources/networking/logical_interconnects.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -30,10 +30,12 @@ from future import standard_library
 
 standard_library.install_aliases()
 
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.exceptions import HPOneViewResourceNotFound
+from hpOneView.resources.resource import (Resource, ResourcePatchMixin, merge_resources,
+                                          ensure_resource_client, unavailable_method)
 
 
-class LogicalInterconnects(object):
+class LogicalInterconnects(ResourcePatchMixin, Resource):
     """
     Logical Interconnects API client.
 
@@ -51,28 +53,40 @@ class LogicalInterconnects(object):
         '200': {"type": "InterconnectSettingsV3"},
         '300': {"type": "InterconnectSettingsV201"},
         '500': {"type": "InterconnectSettingsV201"},
-        '600': {"type": "InterconnectSettingsV201"}
     }
 
     SETTINGS_ETHERNET_DEFAULT_VALUES = {
         '200': {"type": "EthernetInterconnectSettingsV3"},
         '300': {"type": "EthernetInterconnectSettingsV201"},
         '500': {"type": "EthernetInterconnectSettingsV201"},
-        '600': {"type": "EthernetInterconnectSettingsV4"}
+        '600': {"type": "EthernetInterconnectSettingsV4"},
+        '800': {"type": "EthernetInterconnectSettingsV4"}
     }
 
     SETTINGS_TELEMETRY_CONFIG_DEFAULT_VALUES = {
         '200': {"type": "telemetry-configuration"},
         '300': {"type": "telemetry-configuration"},
         '500': {"type": "telemetry-configuration"},
-        '600': {"type": "telemetry-configuration"}
+        '600': {"type": "telemetry-configuration"},
+        '800': {"type": "telemetry-configuration"}
     }
 
-    def __init__(self, con):
-        self._connection = con
-        self._client = ResourceClient(con, self.URI)
+    def __init__(self, connection, data=None):
+        super(LogicalInterconnects, self).__init__(connection, data)
 
-    def get_all(self, start=0, count=-1, filter='', sort=''):
+    def create(self):
+        """Create method is not available for this resource"""
+        unavailable_method()
+
+    def update(self):
+        """Update method is not available for this resource"""
+        unavailable_method()
+
+    def delete(self):
+        """Delete method is not available for this resource"""
+        unavailable_method()
+
+    def get_all(self, start=0, count=-1, sort=''):
         """
         Gets a list of logical interconnects based on optional sorting and filtering and is constrained by start
         and count parameters.
@@ -85,9 +99,6 @@ class LogicalInterconnects(object):
                 The number of resources to return. A count of -1 requests all items.
                 The actual number of items in the response might differ from the requested
                 count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
             sort:
                 The sort order of the returned data set. By default, the sort order is based
                 on create time with the oldest entry first.
@@ -95,19 +106,7 @@ class LogicalInterconnects(object):
         Returns:
             list: A list of logical interconnects.
         """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
-
-    def get(self, id_or_uri):
-        """
-        Gets a logical interconnect by ID or by URI.
-
-        Args:
-            id_or_uri: Can be either the logical interconnect id or the logical interconnect uri.
-
-        Returns:
-            dict: The logical interconnect.
-        """
-        return self._client.get(id_or_uri)
+        return self._helper.get_all(start, count, sort=sort)
 
     def get_by_name(self, name):
         """
@@ -119,11 +118,17 @@ class LogicalInterconnects(object):
         Returns:
             dict: Logical Interconnect.
         """
-        logical_interconnects = self._client.get_all()
+        logical_interconnects = self.get_all()
         result = [x for x in logical_interconnects if x['name'] == name]
-        return result[0] if result else None
+        resource = result[0] if result else None
 
-    def update_compliance(self, id_or_uri, timeout=-1):
+        if resource:
+            resource = self.new(self._connection, resource)
+
+        return resource
+
+    @ensure_resource_client
+    def update_compliance(self, timeout=-1):
         """
         Returns logical interconnects to a consistent state. The current logical interconnect state is
         compared to the associated logical interconnect group.
@@ -135,22 +140,21 @@ class LogicalInterconnects(object):
         existing interconnects from management and/or adding new interconnects for management.
 
         Args:
-            id_or_uri: Can be either the resource ID or the resource URI.
             timeout: Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
                 in OneView; it just stops waiting for its completion.
 
         Returns:
             dict: Logical Interconnect.
         """
-        uri = self._client.build_uri(id_or_uri) + "/compliance"
-        return self._client.update_with_zero_body(uri, timeout=timeout)
+        uri = "{}/compliance".format(self.data["uri"])
+        return self._helper.update(None, uri, timeout=timeout)
 
-    def update_ethernet_settings(self, id_or_uri, configuration, force=False, timeout=-1):
+    @ensure_resource_client
+    def update_ethernet_settings(self, configuration, force=False, timeout=-1):
         """
         Updates the Ethernet interconnect settings for the logical interconnect.
 
         Args:
-            id_or_uri: Can be either the resource ID or the resource URI.
             configuration:  Ethernet interconnect settings.
             force: If set to true, the operation completes despite any problems with network connectivity or errors
                 on the resource itself. The default is false.
@@ -160,15 +164,15 @@ class LogicalInterconnects(object):
         Returns:
             dict: Logical Interconnect.
         """
-        uri = self._client.build_uri(id_or_uri) + "/ethernetSettings"
-        return self._client.update(configuration, uri=uri, force=force, timeout=timeout)
+        uri = "{}/ethernetSettings".format(self.data["uri"])
+        return self._helper.update(configuration, uri=uri, force=force, timeout=timeout)
 
-    def update_internal_networks(self, id_or_uri, network_uri_list, force=False, timeout=-1):
+    @ensure_resource_client
+    def update_internal_networks(self, network_uri_list, force=False, timeout=-1):
         """
         Updates internal networks on the logical interconnect.
 
         Args:
-            id_or_uri: Can be either the resource ID or the resource URI.
             network_uri_list: List of Ethernet network uris.
             force: If set to true, the operation completes despite any problems with network connectivity or errors
                 on the resource itself. The default is false.
@@ -178,30 +182,30 @@ class LogicalInterconnects(object):
         Returns:
             dict: Logical Interconnect.
         """
-        uri = self._client.build_uri(id_or_uri) + "/internalNetworks"
-        return self._client.update(network_uri_list, uri=uri, force=force, timeout=timeout)
+        uri = "{}/internalNetworks".format(self.data["uri"])
+        return self._helper.update(network_uri_list, uri=uri, force=force, timeout=timeout)
 
-    def get_internal_vlans(self, id_or_uri):
+    @ensure_resource_client
+    def get_internal_vlans(self):
         """
         Gets the internal VLAN IDs for the provisioned networks on a logical interconnect.
-
-        Args:
-            id_or_uri: Can be either the logical interconnect group id or the logical interconnect group uri
 
         Returns:
             dict: Collection of URIs
 
         """
-        uri = self._client.build_uri(id_or_uri) + "/internalVlans"
-        return self._client.get_collection(uri)
+        uri = "{}/internalVlans".format(self.data["uri"])
+        response = self._helper.do_get(uri)
 
-    def update_settings(self, id_or_uri, settings, force=False, timeout=-1):
+        return self._helper.get_members(response)
+
+    @ensure_resource_client
+    def update_settings(self, settings, force=False, timeout=-1):
         """
         Updates interconnect settings on the logical interconnect. Changes to interconnect settings are asynchronously
         applied to all managed interconnects.
-
+        (This method is not available from API version 600 onwards)
         Args:
-            id_or_uri: Can be either the resource ID or the resource URI.
             settings: Interconnect settings
             force: If set to true, the operation completes despite any problems with network connectivity or errors
                 on the resource itself. The default is false.
@@ -214,48 +218,49 @@ class LogicalInterconnects(object):
         data = settings.copy()
 
         if 'ethernetSettings' in data:
-            data['ethernetSettings'] = self._client.merge_default_values(data['ethernetSettings'],
-                                                                         self.SETTINGS_ETHERNET_DEFAULT_VALUES)
+            ethernet_default_values = self._get_default_values(self.SETTINGS_ETHERNET_DEFAULT_VALUES)
+            data['ethernetSettings'] = merge_resources(data['ethernetSettings'],
+                                                       ethernet_default_values)
 
-        uri = self._client.build_uri(id_or_uri) + "/settings"
-        return self._client.update(data, uri=uri, force=force, timeout=timeout,
-                                   default_values=self.SETTINGS_DEFAULT_VALUES)
+        uri = "{}/settings".format(self.data["uri"])
+        default_values = self._get_default_values(self.SETTINGS_DEFAULT_VALUES)
+        data = self._helper.update_resource_fields(data, default_values)
 
-    def update_configuration(self, id_or_uri, timeout=-1):
+        return self._helper.update(data, uri=uri, force=force, timeout=timeout)
+
+    @ensure_resource_client
+    def update_configuration(self, timeout=-1):
         """
         Asynchronously applies or re-applies the logical interconnect configuration to all managed interconnects.
 
         Args:
-            id_or_uri: Can be either the resource ID or the resource URI.
             timeout: Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
                 in OneView; it just stops waiting for its completion.
 
         Returns:
             dict: Logical Interconnect.
         """
-        uri = self._client.build_uri(id_or_uri) + "/configuration"
-        return self._client.update_with_zero_body(uri=uri, timeout=timeout)
+        uri = "{}/configuration".format(self.data["uri"])
+        return self._helper.update(None, uri=uri, timeout=timeout)
 
-    def get_snmp_configuration(self, id_or_uri):
+    @ensure_resource_client
+    def get_snmp_configuration(self):
         """
         Gets the SNMP configuration for a logical interconnect.
-
-        Args:
-            id_or_uri: Can be either the logical interconnect group id or the logical interconnect group uri.
 
         Returns:
             dict: SNMP configuration.
         """
-        uri = self._client.build_uri(id_or_uri) + self.SNMP_CONFIGURATION_PATH
-        return self._client.get(uri)
+        uri = "{}{}".format(self.data["uri"], self.SNMP_CONFIGURATION_PATH)
+        return self._helper.do_get(uri)
 
-    def update_snmp_configuration(self, id_or_uri, configuration, timeout=-1):
+    @ensure_resource_client
+    def update_snmp_configuration(self, configuration, timeout=-1):
         """
         Updates the SNMP configuration of a logical interconnect. Changes to the SNMP configuration are asynchronously
         applied to all managed interconnects.
 
         Args:
-            id_or_uri: Can be either the logical interconnect id or the logical interconnect uri.
             configuration: snmp configuration.
 
         Returns:
@@ -265,42 +270,55 @@ class LogicalInterconnects(object):
         if 'type' not in data:
             data['type'] = 'snmp-configuration'
 
-        uri = self._client.build_uri(id_or_uri) + self.SNMP_CONFIGURATION_PATH
-        return self._client.update(data, uri=uri, timeout=timeout)
+        uri = "{}{}".format(self.data["uri"], self.SNMP_CONFIGURATION_PATH)
+        return self._helper.update(data, uri=uri, timeout=timeout)
 
-    def get_unassigned_uplink_ports(self, id_or_uri):
+    @ensure_resource_client
+    def get_unassigned_ports(self):
+        """
+        Gets the collection ports from the member interconnects
+        which are eligible for assignment to an anlyzer port
+
+        Returns:
+            dict: Collection of ports
+        """
+        uri = "{}/unassignedPortsForPortMonitor".format(self.data["uri"])
+        response = self._helper.do_get(uri)
+
+        return self._helper.get_members(response)
+
+    @ensure_resource_client
+    def get_unassigned_uplink_ports(self):
         """
         Gets a collection of uplink ports from the member interconnects which are eligible for assignment to an
         analyzer port. To be eligible, a port must be a valid uplink, must not be a member of an existing uplink set,
         and must not currently be used for stacking.
 
-        Args:
-            id_or_uri: Can be either the logical interconnect group id or the logical interconnect group uri.
-
         Returns:
             dict: Collection of uplink ports.
         """
-        uri = self._client.build_uri(id_or_uri) + "/unassignedUplinkPortsForPortMonitor"
-        return self._client.get_collection(uri)
+        uri = "{}/unassignedUplinkPortsForPortMonitor".format(self.data["uri"])
+        response = self._helper.do_get(uri)
 
-    def get_port_monitor(self, id_or_uri):
+        return self._helper.get_members(response)
+
+    @ensure_resource_client
+    def get_port_monitor(self):
         """
         Gets the port monitor configuration of a logical interconnect.
 
-        Args:
-            id_or_uri: Can be either the logical interconnect id or the logical interconnect uri.
         Returns:
             dict: The Logical Interconnect.
         """
-        uri = self._client.build_uri(id_or_uri) + self.PORT_MONITOR_PATH
-        return self._client.get(uri)
+        uri = "{}{}".format(self.data["uri"], self.PORT_MONITOR_PATH)
+        return self._helper.do_get(uri)
 
-    def update_port_monitor(self, id_or_uri, resource, timeout=-1):
+    @ensure_resource_client
+    def update_port_monitor(self, resource, timeout=-1):
         """
         Updates the port monitor configuration of a logical interconnect.
 
         Args:
-            id_or_uri: Can be either the logical interconnect id or the logical interconnect uri.
             resource: Port monitor configuration.
 
         Returns:
@@ -310,21 +328,8 @@ class LogicalInterconnects(object):
         if 'type' not in data:
             data['type'] = 'port-monitor'
 
-        uri = self._client.build_uri(id_or_uri) + self.PORT_MONITOR_PATH
-        return self._client.update(data, uri=uri, timeout=timeout)
-
-    def get_telemetry_configuration(self, telemetry_configuration_uri):
-        """
-        Gets the telemetry configuration of a logical interconnect.
-
-        Args:
-            telemetry_configuration_uri: Telemetry Configuration URI.
-
-        Returns:
-            dict: Telemetry configuration.
-
-        """
-        return self._client.get(telemetry_configuration_uri)
+        uri = "{}{}".format(self.data["uri"], self.PORT_MONITOR_PATH)
+        return self._helper.update(data, uri=uri, timeout=timeout)
 
     def create_interconnect(self, location_entries, timeout=-1):
         """
@@ -343,7 +348,7 @@ class LogicalInterconnects(object):
         Returns:
             dict: Created interconnect.
         """
-        return self._client.create(location_entries, uri=self.locations_uri, timeout=timeout)
+        return self._helper.create(location_entries, uri=self.locations_uri, timeout=timeout)
 
     def delete_interconnect(self, enclosure_uri, bay, timeout=-1):
         """
@@ -366,22 +371,21 @@ class LogicalInterconnects(object):
         uri = "{path}?location=Enclosure:{enclosure_uri},Bay:{bay}".format(path=self.LOCATIONS_PATH,
                                                                            enclosure_uri=enclosure_uri,
                                                                            bay=bay)
-        return self._client.delete(uri, timeout=timeout)
+        return self._helper.delete(uri, timeout=timeout)
 
-    def get_firmware(self, id_or_uri):
+    @ensure_resource_client
+    def get_firmware(self):
         """
         Gets the installed firmware for a logical interconnect.
-
-        Args:
-            id_or_uri: Can be either the logical interconnect id or the logical interconnect uri.
 
         Returns:
             dict: LIFirmware.
         """
-        firmware_uri = self._client.build_subresource_uri(id_or_uri, subresource_path=self.FIRMWARE_PATH)
-        return self._client.get(firmware_uri)
+        firmware_uri = self._helper.build_subresource_uri(self.data["uri"], subresource_path=self.FIRMWARE_PATH)
+        return self._helper.do_get(firmware_uri)
 
-    def install_firmware(self, firmware_information, id_or_uri):
+    @ensure_resource_client
+    def install_firmware(self, firmware_information):
         """
         Installs firmware to a logical interconnect. The three operations that are supported for the firmware
         update are Stage (uploads firmware to the interconnect), Activate (installs firmware on the interconnect),
@@ -389,22 +393,20 @@ class LogicalInterconnects(object):
 
         Args:
             firmware_information: Options to install firmware to a logical interconnect.
-            id_or_uri: Can be either the logical interconnect id or the logical interconnect uri.
 
         Returns:
             dict
         """
-        firmware_uri = self._client.build_subresource_uri(id_or_uri, subresource_path=self.FIRMWARE_PATH)
-        return self._client.update(firmware_information, firmware_uri)
+        firmware_uri = self._helper.build_subresource_uri(self.data["uri"], subresource_path=self.FIRMWARE_PATH)
+        return self._helper.update(firmware_information, firmware_uri)
 
-    def get_forwarding_information_base(self, id_or_uri, filter=''):
+    @ensure_resource_client
+    def get_forwarding_information_base(self, filter=''):
         """
         Gets the forwarding information base data for a logical interconnect. A maximum of 100 entries is returned.
         Optional filtering criteria might be specified.
 
         Args:
-            id_or_uri:
-                Can be either the logical interconnect id or the logical interconnect uri.
             filter (list or str):
                 Filtering criteria may be specified using supported attributes: interconnectUri, macAddress,
                 internalVlan, externalVlan, and supported relation = (Equals). macAddress is 12 hexadecimal digits with
@@ -414,46 +416,41 @@ class LogicalInterconnects(object):
         Returns:
             list: A set of interconnect MAC address entries.
         """
-        uri = self._client.build_uri(id_or_uri) + self.FORWARDING_INFORMATION_PATH
-        return self._client.get_collection(uri, filter=filter)
+        uri = "{}{}".format(self.data["uri"], self.FORWARDING_INFORMATION_PATH)
+        return self._helper.get_collection(uri, filter=filter)
 
-    def create_forwarding_information_base(self, id_or_uri, timeout=-1):
+    @ensure_resource_client
+    def create_forwarding_information_base(self, timeout=-1):
         """
         Generates the forwarding information base dump file for a logical interconnect.
 
         Args:
-            id_or_uri:
-                Can be either the logical interconnect id or the logical interconnect uri.
             timeout:
                 Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation in
                 OneView, just stops waiting for its completion.
 
         Returns: Interconnect Forwarding Information Base DataInfo.
         """
-        uri = self._client.build_uri(id_or_uri) + self.FORWARDING_INFORMATION_PATH
-        return self._client.create_with_zero_body(uri=uri, timeout=timeout)
+        uri = "{}{}".format(self.data["uri"], self.FORWARDING_INFORMATION_PATH)
+        return self._helper.do_post(uri, None, timeout, None)
 
-    def get_qos_aggregated_configuration(self, id_or_uri):
+    @ensure_resource_client
+    def get_qos_aggregated_configuration(self):
         """
         Gets the QoS aggregated configuration for the logical interconnect.
-
-        Args:
-            id_or_uri:
-                Can be either the logical interconnect id or the logical interconnect uri.
 
         Returns:
             dict: QoS Configuration.
         """
-        uri = self._client.build_uri(id_or_uri) + self.QOS_AGGREGATED_CONFIGURATION
-        return self._client.get(uri)
+        uri = "{}{}".format(self.data["uri"], self.QOS_AGGREGATED_CONFIGURATION)
+        return self._helper.do_get(uri)
 
-    def update_qos_aggregated_configuration(self, id_or_uri, qos_configuration, timeout=-1):
+    @ensure_resource_client
+    def update_qos_aggregated_configuration(self, qos_configuration, timeout=-1):
         """
         Updates the QoS aggregated configuration for the logical interconnect.
 
         Args:
-            id_or_uri:
-                Can be either the logical interconnect id or the logical interconnect uri.
             qos_configuration:
                 QOS configuration.
             timeout:
@@ -463,21 +460,36 @@ class LogicalInterconnects(object):
         Returns:
             dict: Logical Interconnect.
         """
-        uri = self._client.build_uri(id_or_uri) + self.QOS_AGGREGATED_CONFIGURATION
-        return self._client.update(qos_configuration, uri=uri, timeout=timeout)
+        uri = "{}{}".format(self.data["uri"], self.QOS_AGGREGATED_CONFIGURATION)
+        return self._helper.update(qos_configuration, uri=uri, timeout=timeout)
 
-    def update_telemetry_configurations(self, tc_id_or_uri, configuration, id_or_uri=None, timeout=-1):
+    def _get_telemetry_configuration_uri(self):
+        telemetry_conf = self.data.get("telemetryConfiguration", {})
+        if not telemetry_conf.get("uri"):
+            raise HPOneViewResourceNotFound("Telemetry configuration uri is not available")
+        return telemetry_conf["uri"]
+
+    @ensure_resource_client
+    def get_telemetry_configuration(self):
+        """
+        Gets the telemetry configuration of a logical interconnect.
+
+        Returns:
+            dict: Telemetry configuration.
+
+        """
+        telemetry_conf_uri = self._get_telemetry_configuration_uri()
+        return self._helper.do_get(telemetry_conf_uri)
+
+    @ensure_resource_client
+    def update_telemetry_configurations(self, configuration, timeout=-1):
         """
         Updates the telemetry configuration of a logical interconnect. Changes to the telemetry configuration are
         asynchronously applied to all managed interconnects.
 
         Args:
-            tc_id_or_uri:
-                Can be either the telemetry configuration id or uri.
             configuration:
                 The telemetry configuration for the logical interconnect.
-            id_or_uri:
-                Can be either the logical interconnect id or uri.
             timeout:
                 Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation in
                 OneView, just stops waiting for its completion.
@@ -485,39 +497,19 @@ class LogicalInterconnects(object):
         Returns:
             dict: The Logical Interconnect.
         """
-        uri = self._client.build_subresource_uri(id_or_uri, tc_id_or_uri, 'telemetry-configurations')
-        return self._client.update(configuration, uri=uri, timeout=timeout,
-                                   default_values=self.SETTINGS_TELEMETRY_CONFIG_DEFAULT_VALUES)
+        telemetry_conf_uri = self._get_telemetry_configuration_uri()
+        default_values = self._get_default_values(self.SETTINGS_TELEMETRY_CONFIG_DEFAULT_VALUES)
+        configuration = self._helper.update_resource_fields(configuration, default_values)
 
-    def get_ethernet_settings(self, id_or_uri):
+        return self._helper.update(configuration, uri=telemetry_conf_uri, timeout=timeout)
+
+    @ensure_resource_client
+    def get_ethernet_settings(self):
         """
         Gets the Ethernet interconnect settings for the Logical Interconnect.
-
-        Args:
-            id_or_uri:
-                Can be either the logical interconnect id or uri.
 
         Returns:
             dict: Ethernet Interconnect Settings
         """
-        uri = self._client.build_uri(id_or_uri) + '/ethernetSettings'
-        return self._client.get(uri)
-
-    def patch(self, id_or_uri, operation, path, value, timeout=-1):
-        """
-        Uses the PATCH to update a resource for a given logical interconnect.
-
-        Only one operation can be performed in each PATCH call.
-
-        Args:
-            id_or_uri: Can be either the resource ID or the resource URI.
-            operation: Patch operation
-            path: Path
-            value: Value
-            timeout: Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView; it just stops waiting for its completion.
-
-        Returns:
-            dict: Updated resource.
-        """
-        return self._client.patch(id_or_uri, operation, path, value, timeout=timeout)
+        uri = "{}/ethernetSettings".format(self.data["uri"])
+        return self._helper.do_get(uri)